fix(context-cache): widen the demand window past the replay distance
Public value accumulates only over the bits of a committed demand mask, and
the window behind that mask held 32 requests. A workload that replays a few
dozen distinct prefixes twice evicts the first pass's demand record before the
second pass asks for the same key, so every candidate reads as a single reuse
domain, the repeated gate never fires and a shared publication can never show
positive value: the prefix is re-prefilled every time. The window now holds 64
requests, beyond the replay distance of that workload.

+++ src/runtime/engine/context_cache/materialization_planner.h
@@ -25,7 +25,7 @@ struct MaterializationCheckpointPolicy {
     RetentionClass retention_class     = RetentionClass::RecentPrivate;
     std::uint64_t selected_hit_count   = 0;
     std::uint64_t last_hit_epoch       = 0;
-    std::uint32_t demand_mask          = 0;
+    std::uint64_t demand_mask          = 0;
     std::uint64_t rebuild_ns           = 0;
     std::uint64_t baseline_recovery_ns = 0;
 };
